inference/requirements_gpu.txt
# See: https://github.com/ovh/ai-training-examples/blob/main/jobs/flask/object-detection-yolov5-app/requirements.txt

yolox==0.3.0

onnxruntime-gpu

inventa==0.0a15

prometheus-client